private profile & delete account

+++ src/controllers/RelationshipController.js
@@ -10,11 +10,12 @@ const ApiError = require('../utils/ApiError');
 const userRepository = AppDataSource.getRepository(User);
 const relationshipRepository = AppDataSource.getRepository(Relationship);
 const friendRequestRepository = AppDataSource.getRepository(FriendRequest);
-const relationshipTypeRepository = AppDataSource.getRepository(RelationshipType);
+const relationshipTypeRepository =
+  AppDataSource.getRepository(RelationshipType);
 const notificationsRepository = AppDataSource.getRepository(Notifications);
 const notificationTypeRepository =
   AppDataSource.getRepository(NotificationType);
-  
+
 class RelationshipController {
   // get common friends
   async commonFriends(userId1, userId2) {
@@ -253,21 +254,24 @@ class RelationshipController {
     const { relationship, receiverId } = req.body;
     const { id } = req.userToken;
 
-    const relationshipTypeFriend = await relationshipTypeRepository.findOne({where: {
-      name: "Bạn bè"
-    }})
-    
+    const relationshipTypeFriend = await relationshipTypeRepository.findOne({
+      where: {
+        name: 'Bạn bè',
+      },
+    });
+
     if (!relationship || relationship === relationshipTypeFriend?.id) {
       const newFriendRequest = await friendRequestRepository.save({
         senderId: id,
         receiverId: receiverId,
       });
 
-      const notificationTypeFriendRequest = await notificationTypeRepository.findOne({
-        where: {
-          name: 'friend request'
-        }
-      })
+      const notificationTypeFriendRequest =
+        await notificationTypeRepository.findOne({
+          where: {
+            name: 'friend request',
+          },
+        });
 
       const userInfo = await userRepository.findOne({
         where: { id: id },
@@ -287,8 +291,8 @@ class RelationshipController {
         senderId: id,
         type: notificationTypeFriendRequest?.id,
         relatedId: newFriendRequest?.id,
-        content: `<b>${userInfo?.lastName} ${userInfo?.firstName}</b> vừa gửi lời mời kết bạn cho bạn`
-      }) 
+        content: `<b>${userInfo?.lastName} ${userInfo?.firstName}</b> vừa gửi lời mời kết bạn cho bạn`,
+      });
 
       const commonFriends = this.commonFriends(id, receiverId);
 
@@ -298,7 +302,10 @@ class RelationshipController {
         numberOfCommonFriends: commonFriends?.length,
       });
 
-      io.to(`user-${receiverId}`).emit('notificationNewFriendRequest', notificationFriendRequest)
+      io.to(`user-${receiverId}`).emit(
+        'notificationNewFriendRequest',
+        notificationFriendRequest
+      );
     } else {
       await relationshipRepository.save({
         user1: id,
@@ -362,19 +369,20 @@ class RelationshipController {
       },
     });
 
-    const notificationTypeFriendRequest = notificationTypeRepository.findOne({
-      where: {
-        name: 'friend request'
-      }
-    })
+    const notificationTypeFriendRequest =
+      await notificationTypeRepository.findOne({
+        where: {
+          name: 'friend request',
+        },
+      });
 
     const notificationFriendRequest = await notificationsRepository.findOne({
       where: {
         senderId,
         userId: id,
-        type: notificationTypeFriendRequest.id
-      }
-    })
+        type: notificationTypeFriendRequest.id,
+      },
+    });
 
     if (friendRequest && notificationFriendRequest) {
       await friendRequestRepository.remove(friendRequest);
@@ -408,21 +416,26 @@ class RelationshipController {
       },
     });
 
-    const notificationTypeFriendRequest = notificationTypeRepository.findOne({
-      where: {
-        name: 'friend request'
-      }
-    })
+    const notificationTypeFriendRequest =
+      await notificationTypeRepository.findOne({
+        where: {
+          name: 'friend request',
+        },
+      });
 
     const notificationFriendRequest = await notificationsRepository.findOne({
       where: {
         senderId: friendId,
         userId: id,
-        type: notificationTypeFriendRequest.id
-      }
-    })
+        type: notificationTypeFriendRequest.id,
+      },
+    });
 
-    if (!existedRelationship && existedFriendRequest && notificationFriendRequest) {
+    if (
+      !existedRelationship &&
+      existedFriendRequest &&
+      notificationFriendRequest
+    ) {
       await relationshipRepository.save({
         user1: id,
         user2: friendId,
@@ -529,8 +542,24 @@ class RelationshipController {
       },
     });
 
-    if (friendRequest) {
+    const notificationTypeFriendRequest =
+      await notificationTypeRepository.findOne({
+        where: {
+          name: 'friend request',
+        },
+      });
+
+    const notificationFriendRequest = await notificationsRepository.findOne({
+      where: {
+        senderId: friendId,
+        userId: id,
+        type: notificationTypeFriendRequest.id,
+      },
+    });
+
+    if (friendRequest && notificationFriendRequest) {
       await friendRequestRepository.remove(friendRequest);
+      await notificationsRepository.remove(notificationFriendRequest);
 
       io.to(`user-${receiverId}`).emit('cancelFriendRequest', id);
